Vendor all Bundler installments.
If caching is turned on, Travis vendors only root installment. However,
this also makes Bundler to look into `vendor/bundle` in dummy apps,
while we `bundle install` their dependencies system wide. That make
a build to fail.

Therefore, in order Travis build to be finished and to cache app's gems
(not just Torba's gems), we have to provide `--path` in all invocations
of `bundle install`.

For local development this is not needed (actually, using system wide
gems there would probably be a faster option).